Refactored custom form element ordering
The UI now has a hidden position field for each form field. I've changed the
drag/drop ordering mechanism to simply re-order these elements rather than
make an AJAX call to update the ordering in the database.

This means reordering ONLY happens on when the user hits the save button. It
also fixes the problem where the ordering was broken by a user hitting the
save button.